Create MESG Overview
Here is a draft of the new Overview page for docs. I would like to add more information in the "Developer language SDKs" section to describe the languages that can be used. I want to make it clear that MESG is language-flexible, and have that be described in a technical way. Otherwise, we could remove this section.